Use gameRoom.players instead of tokens
